refactor: modularize Rust value adapters
Centralize ambiguous type identity resolution so qualified-name lookup and
adapter selection share one policy. Split layout validation and tests by
adapter family without changing fallback or capture-plan behavior.
